toys/fullstack/strawberry_graphql/requirements.txt
strawberry-graphql[debug-server]~=0.213